Adding support to organize text level files
When writing a level file in text format, it needs to be in a form a user can
edit easily. So I added endSection, endStructure to the NumberOutputStream
class. These methods only have effect when writing text, and they cause extra
new-lines to be written to the file.

Also made it so spaces are only inserted when necessary, so they don't trail at
the end or start of lines.