Enabling level saving using the Java LevelMaker
The old level maker is no longer needed.

Checked to make sure the level save result is the same as the existing levels.
It is, except for whitespace and the palette size. Whitespace is now easier to
read, and the palette size is printed as "0" instead of "256", which is how the
binary level format works.