Making items' black pixels transparent
Before, LevelMaker was showing the black of each image as actual black, so the
images don't look like they will in the game. Added logic to process each image
to turn black pixels to transparent pixels.

The 1turners.bmp image had slightly off-black in its palette instead of pure
black, so it didn't get processed properly. So I changed the color in its
palette from 040204 to 000000, so the black pixels will be made transparent.